requirements.txt
django==3.1.5
djangorestframework==3.12.2
djangorestframework_simplejwt==4.6.0
pytest-django==4.1.0
pytest-cov==2.8.1
pytest==5.4.0
django-heroku==0.3.1
requests==2.25.1